+++ src/elements/GeoList/GeoListItem.vue
@@ -1,22 +1,21 @@
-<template functional>
+<template>
   <component
-    :is="props.wrapperTag"
-    :ref="data.ref"
-    v-bind="data.attrs"
-    :class="[data.class, data.staticClass, {
+    :is="wrapperTag"
+    v-bind="$attrs"
+    :class="[{
       'geo-list-item': true,
-      'geo-list-item--disabled': props.disabled
+      'geo-list-item--disabled': disabled
     }]"
-    v-on="listeners"
+    v-on="computedListeners"
   >
     <div class="geo-list-item__label-and-accessory-container">
       <div class="geo-list-item__icon-and-label">
         <div
-          v-if="props.icon"
+          v-if="icon"
           class="geo-list-item__icon-and-label__icon-container"
         >
           <font-awesome-icon
-            :icon="props.icon"
+            :icon="icon"
             class="geo-list-item__icon-and-label__icon-container__icon"
             aria-hidden
             fixed-width
@@ -39,7 +38,7 @@
       v-if="$slots.description"
       :class="{
         'geo-list-item__description': true,
-        'geo-list-item__description--spaced': !!props.icon
+        'geo-list-item__description--spaced': !!icon
       }"
     >
       <!-- @slot Use this slot to add a description for the item -->
@@ -49,6 +48,8 @@
 </template>
 
 <script>
+import extendedListenersMixin from '@/mixins/extendedListenersMixin'
+
 /**
  * `GeoListItem` is a component designed to build vertical lists which fit
  * properly in a `GeoBorderedBox`.
@@ -60,6 +61,7 @@ export default {
   name: 'GeoListItem',
   status: 'ready',
   release: '4.0.0',
+  mixins: [extendedListenersMixin],
   props: {
     /**
      * Optional Font Awesome 5 icon to be displayed next to the entry's label,