+++ test/unit/specs/elements/GeoList/GeoListItem.spec.js
@@ -1,4 +1,4 @@
-import { mount } from '@vue/test-utils'
+import { shallowMount } from '@vue/test-utils'
 import GeoListItem from '@/elements/GeoList/GeoListItem.vue'
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
 import { library } from '@fortawesome/fontawesome-svg-core'
@@ -8,12 +8,11 @@ library.add(fas)
 
 describe('GeoListItem', () => {
   it('Should render a <div> wrapper by default', () => {
-    const wrapper = mount(GeoListItem)
-    expect(wrapper.find('div.geo-list-item').exists()).toBe(true)
+    expect(getShallowWrapper().find('div.geo-list-item').exists()).toBe(true)
   })
 
   it('Should render a <label> wrapper if wrapperTag is provided', () => {
-    const wrapper = mount(GeoListItem, {
+    const wrapper = getShallowWrapper({
       propsData: {
         wrapperTag: 'label'
       }
@@ -22,42 +21,8 @@ describe('GeoListItem', () => {
     expect(wrapper.find('label.geo-list-item').exists()).toBe(true)
   })
 
-  it('Should render with the provided static class', () => {
-    const wrapper = mount(GeoListItem, {
-      context: {
-        staticClass: 'my-static-class'
-      }
-    })
-    expect(wrapper.find('div.geo-list-item.my-static-class').exists()).toBe(true)
-  })
-
-  it('Should render with the provided dynamic class', () => {
-    const wrapper = mount(GeoListItem, {
-      context: {
-        class: {
-          'my-dynamic-class': true,
-          'my-dynamic-class-false': false
-        }
-      }
-    })
-    expect(wrapper.find('div.geo-list-item.my-dynamic-class').exists()).toBe(true)
-    expect(wrapper.find('div.geo-list-item.my-dynamic-class-false').exists()).toBe(false)
-  })
-
-  it('Should render with the provided dynamic and static classes', () => {
-    const wrapper = mount(GeoListItem, {
-      context: {
-        class: {
-          'my-dynamic-class': true
-        },
-        staticClass: 'my-static-class'
-      }
-    })
-    expect(wrapper.find('div.geo-list-item.my-static-class.my-dynamic-class').exists()).toBe(true)
-  })
-
   it('Should render default slot', function () {
-    const wrapper = mount(GeoListItem, {
+    const wrapper = getShallowWrapper({
       slots: {
         default: ['<span class="my-demo-content">Just some unique demo content</span>']
       }
@@ -67,7 +32,7 @@ describe('GeoListItem', () => {
   })
 
   it('Should render trailingAccessoryItem slot', function () {
-    const wrapper = mount(GeoListItem, {
+    const wrapper = getShallowWrapper({
       slots: {
         default: ['<span class="my-demo-content">Just some unique demo content</span>'],
         trailingAccessoryItem: ['<span class="my-demo-accessory">Just some accessory item</span>']
@@ -78,14 +43,12 @@ describe('GeoListItem', () => {
   })
 
   it('Should render icon', function () {
-    const wrapper = mount(GeoListItem, {
+    const wrapper = getShallowWrapper({
       stubs: {
         'font-awesome-icon': FontAwesomeIcon
       },
-      context: {
-        props: {
-          icon: ['fas', 'user']
-        }
+      propsData: {
+        icon: ['fas', 'user']
       }
     })
 
@@ -93,7 +56,7 @@ describe('GeoListItem', () => {
   })
 
   it('Should render description', function () {
-    const wrapper = mount(GeoListItem, {
+    const wrapper = getShallowWrapper({
       slots: {
         description: ['<span class="my-demo-description">Description</span>']
       }
@@ -104,22 +67,37 @@ describe('GeoListItem', () => {
 
   it('Should emit click event', function () {
     const clickListener = jest.fn()
-    const wrapper = mount(GeoListItem, {
+    const wrapper = getShallowWrapper({
       slots: {
         default: ['Just some unique demo content']
       },
-      context: {
-        on: {
-          click: clickListener
-        }
+      listeners: {
+        click: clickListener
       }
     })
     wrapper.find('.geo-list-item').trigger('click')
     expect(clickListener).toHaveBeenCalled()
   })
 
+  it('Should not emit click event when is disabled', function () {
+    const clickListener = jest.fn()
+    const wrapper = getShallowWrapper({
+      slots: {
+        default: ['Just some unique demo content']
+      },
+      listeners: {
+        click: clickListener
+      },
+      propsData: {
+        disabled: true
+      }
+    })
+    wrapper.find('.geo-list-item').trigger('click')
+    expect(clickListener).not.toHaveBeenCalled()
+  })
+
   it('Should include disabled suffix when it is disabled', function () {
-    const wrapper = mount(GeoListItem, {
+    const wrapper = getShallowWrapper({
       propsData: {
         disabled: true
       }
@@ -128,3 +106,7 @@ describe('GeoListItem', () => {
     expect(wrapper.find('.geo-list-item--disabled').exists()).toBe(true)
   })
 })
+
+function getShallowWrapper (options) {
+  return shallowMount(GeoListItem, options)
+}